+++ src/Compute/Compute/Manual/VirtualMachineScaleSetCreateOrUpdateMethod.cs
@@ -690,6 +690,14 @@ namespace Microsoft.Azure.Commands.Compute.Automation
             
             var parameters = new Parameters(this, client);
 
+            if (shouldGuestAttestationExtBeInstalledSimple()
+                && !this.IsParameterBound(c => c.SystemAssignedIdentity)
+                && !this.IsParameterBound(c => c.UserAssignedIdentity)
+                    )
+            {
+                this.SystemAssignedIdentity = true;
+            }
+
             if (parameters?.ImageAndOsType?.Image?.Version?.ToLower() != "latest")
             {
                 WriteWarning("You are deploying VMSS pinned to a specific image version from Azure Marketplace. \n" +
@@ -725,6 +733,111 @@ namespace Microsoft.Azure.Commands.Compute.Automation
                     port, 
                     range);
                 asyncCmdlet.WriteObject(psObject);
+            }
+
+            if (shouldGuestAttestationExtBeInstalledSimple())
+            {
+                string extensionNameGA = "GuestAttestation";
+                var extensionDirect = new VirtualMachineScaleSetExtension();
+
+                if (result.VirtualMachineProfile?.OsProfile != null)
+                {
+                    if (result.VirtualMachineProfile.OsProfile.LinuxConfiguration != null)
+                    {
+
+                        extensionDirect.Name = extensionNameGA;
+                        extensionDirect.Publisher = "Microsoft.Azure.Security.LinuxAttestation";
+                        extensionDirect.Type1 = extensionNameGA;
+                        extensionDirect.TypeHandlerVersion = "1.0";
+                        extensionDirect.EnableAutomaticUpgrade = true;
+                    }
+                    else
+                    {
+
+                        extensionDirect.Name = extensionNameGA;
+                        extensionDirect.Publisher = "Microsoft.Azure.Security.WindowsAttestation";
+                        extensionDirect.Type1 = extensionNameGA;
+                        extensionDirect.TypeHandlerVersion = "1.0";
+                        extensionDirect.EnableAutomaticUpgrade = true;
+                    }
+                }
+
+                VirtualMachineScaleSetUpdate parametersupdate = new VirtualMachineScaleSetUpdate();
+                parametersupdate.VirtualMachineProfile = new VirtualMachineScaleSetUpdateVMProfile();
+                parametersupdate.VirtualMachineProfile.ExtensionProfile = new VirtualMachineScaleSetExtensionProfile();
+                parametersupdate.VirtualMachineProfile.ExtensionProfile.Extensions = new List<VirtualMachineScaleSetExtension>();
+                parametersupdate.VirtualMachineProfile.ExtensionProfile.Extensions.Add(extensionDirect);
+                result = VirtualMachineScaleSetsClient.Update(this.ResourceGroupName, this.VMScaleSetName, parametersupdate);
+
+                var vmssVmExtParams = new VirtualMachineScaleSetVMExtension();
+                var resultVmssVm = VirtualMachineScaleSetVMsClient.List(this.ResourceGroupName, this.VMScaleSetName);
+                var resultList = resultVmssVm.ToList();
+                var nextPageLink = resultVmssVm.NextPageLink;
+                while (!string.IsNullOrEmpty(nextPageLink))
+                {
+                    var pageResult = VirtualMachineScaleSetVMsClient.ListNext(nextPageLink);
+                    foreach (var pageItem in pageResult)
+                    {
+                        resultList.Add(pageItem);
+                    }
+                    nextPageLink = pageResult.NextPageLink;
+                }
+
+                foreach (var currentVmssVm in resultList)
+                {
+                    if (currentVmssVm.StorageProfile != null &&
+                        currentVmssVm.StorageProfile.OsDisk != null)
+                    {
+                        if (currentVmssVm.StorageProfile.OsDisk.OsType == OperatingSystemTypes.Linux)
+                        {
+                            vmssVmExtParams = new VirtualMachineScaleSetVMExtension
+                            {
+                                Publisher = "Microsoft.Azure.Security.LinuxAttestation",
+                                Type1 = extensionNameGA,
+                                TypeHandlerVersion = "1.0",
+                                EnableAutomaticUpgrade = true
+                            };
+                        }
+                        else
+                        {
+
+                            vmssVmExtParams = new VirtualMachineScaleSetVMExtension
+                            {
+                                Publisher = "Microsoft.Azure.Security.WindowsAttestation",
+                                Type1 = extensionNameGA,
+                                TypeHandlerVersion = "1.0",
+                                EnableAutomaticUpgrade = true
+                            };
+                        }
+                        var opt = this.VirtualMachineScaleSetVMExtensionsClient.CreateOrUpdateWithHttpMessagesAsync(this.ResourceGroupName, this.VMScaleSetName, currentVmssVm.InstanceId, extensionNameGA, vmssVmExtParams);
+                    }
+
+                }
+            }
+
+        }
+
+        /// <summary>
+        /// Check to see if the Guest Attestation extension should be installed and Identity set to SystemAssigned.
+        /// Requirements for this scenario to be true:
+        /// 1) DisableIntegrityMonitoring is not true.
+        /// 2) SecurityType is TrustedLaunch.
+        /// 3) SecureBootEnabled is true.
+        /// 4) VTpmEnabled is true.
+        /// </summary>
+        /// <returns></returns>
+        private bool shouldGuestAttestationExtBeInstalledSimple()
+        {
+            if (this.DisableIntegrityMonitoring != true &&
+                    this.SecurityType?.ToLower() == ConstantValues.TrustedLaunchSecurityType &&
+                    this.EnableSecureBoot == true &&
+                    this.EnableVtpm == true)
+            {
+                return true;
+            }
+            else
+            {
+                return false;
             }
         }