+++ src/Compute/Compute/ChangeLog.md
@@ -25,7 +25,7 @@
 * [Breaking change] Removed unversioned and outdated linux image aliases of `CentOS`, `RHEL`, `UbuntuLTS` and `Debian`.
 * [Breaking Change] Added defaulting logic for VM and VMSS creation to set SecurityType to TrustedLaunch and SecureBootEnabled and VTpmEnalbed to true when those are not set by the user.
 * [Breaking Change] Added defaulting logic for Disk creation to default to TrustedLaunch when able. Allows the user to turn this off by setting the SecurityType to Standard.
-* [Breaking Change] Removed the GuestAttestation extension installation by default for VM and VMSS resources. Removed the associated `DisableIntegrityMonitoring` parameter from `New-AzVM` and `New-AzVmss` cmdlets.
+* Set the `DisableIntegrityMonitoring` parameter to default to true for `New-AzVM` and `New-AzVmss` cmdlets.
 * Added new parameters `-ElasticSanResourceId` and `-OptimizedForFrequentAttach` to `New-AzDiskConfig` cmdlet.
 * Added new parameter `-ElasticSanResourceId` to `New-AzSnapshotConfig` cmdlet.